R662: consolidate track UI components
